# AdministadorTareasAPI/AdministadorTareasAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AdministadorTareasAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Colaborador/Lista": {
      "get": {
        "tags": [
          "Colaborador"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarea/Lista": {
      "get": {
        "tags": [
          "Tarea"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarea/Obtener/{idTarea}": {
      "get": {
        "tags": [
          "Tarea"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idTarea",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarea/Guardar": {
      "post": {
        "tags": [
          "Tarea"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tarea"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarea/Editar": {
      "put": {
        "tags": [
          "Tarea"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tarea"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarea/Eliminar/{idTarea}": {
      "delete": {
        "tags": [
          "Tarea"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idTarea",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Tarea": {
      "type": "object",
      "properties": {
        "idTarea": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "prioridad": {
          "type": "string"
        },
        "fechaInicio": {
          "format": "date-time",
          "type": "string"
        },
        "fechaFinal": {
          "format": "date-time",
          "type": "string"
        },
        "idColaborador": {
          "format": "int32",
          "type": "integer"
        },
        "nota": {
          "type": "string"
        }
      }
    }
  }
}